+++ python/pyspark/pandas/tests/connect/data_type_ops/test_parity_num_ops_fractional_ext.py
@@ -15,14 +15,16 @@
 # limitations under the License.
 #
 
-from pyspark.pandas.tests.data_type_ops.test_num_ops import NumOpsTestsMixin
+from pyspark.pandas.tests.data_type_ops.test_num_ops_fractional_ext import (
+    FractionalExtensionOpsTestsMixin,
+)
 from pyspark.pandas.tests.data_type_ops.testing_utils import OpsTestBase
 from pyspark.testing.pandasutils import PandasOnSparkTestUtils
 from pyspark.testing.connectutils import ReusedConnectTestCase
 
 
-class NumOpsParityTests(
-    NumOpsTestsMixin,
+class FractionalExtensionOpsParityTests(
+    FractionalExtensionOpsTestsMixin,
     PandasOnSparkTestUtils,
     OpsTestBase,
     ReusedConnectTestCase,
