+++ python/pyspark/pandas/tests/connect/data_type_ops/test_parity_num_ops_fractional_ext_astype_cmp.py
@@ -15,14 +15,16 @@
 # limitations under the License.
 #
 
-from pyspark.pandas.tests.data_type_ops.test_num_ops import NumOpsTestsMixin
+from pyspark.pandas.tests.data_type_ops.test_num_ops_fractional_ext_astype_cmp import (
+    FractionalExtensionAstypeCmpOpsTestsMixin,
+)
 from pyspark.pandas.tests.data_type_ops.testing_utils import OpsTestBase
 from pyspark.testing.pandasutils import PandasOnSparkTestUtils
 from pyspark.testing.connectutils import ReusedConnectTestCase
 
 
-class NumOpsParityTests(
-    NumOpsTestsMixin,
+class FractionalExtensionAstypeCmpOpsParityTests(
+    FractionalExtensionAstypeCmpOpsTestsMixin,
     PandasOnSparkTestUtils,
     OpsTestBase,
     ReusedConnectTestCase,
