+++ python/pyspark/pandas/tests/data_type_ops/test_num_ops.py
@@ -15,8 +15,6 @@
 # limitations under the License.
 #
 
-import unittest
-
 import pandas as pd
 import numpy as np
 
@@ -25,10 +23,6 @@ from pyspark.pandas.config import option_context
 from pyspark.testing.pandasutils import PandasOnSparkTestCase
 from pyspark.testing.utils import is_ansi_mode_test
 from pyspark.pandas.tests.data_type_ops.testing_utils import OpsTestBase
-from pyspark.pandas.typedef.typehints import (
-    extension_dtypes_available,
-    extension_float_dtypes_available,
-)
 from pyspark.sql.types import DecimalType, IntegralType
 
 
@@ -199,256 +193,8 @@ class NumOpsTestsMixin:
                 self.assert_eq(pser >= pser, psser >= psser)
 
 
-@unittest.skipIf(not extension_dtypes_available, "pandas extension dtypes are not available")
-class IntegralExtensionOpsTestsMixin:
-    @property
-    def intergral_extension_psers(self):
-        return [pd.Series([1, 2, 3, None], dtype=dtype) for dtype in self.integral_extension_dtypes]
-
-    @property
-    def intergral_extension_pssers(self):
-        return [ps.from_pandas(pser) for pser in self.intergral_extension_psers]
-
-    @property
-    def intergral_extension_pser_psser_pairs(self):
-        return zip(self.intergral_extension_psers, self.intergral_extension_pssers)
-
-    def test_from_to_pandas(self):
-        for pser, psser in self.intergral_extension_pser_psser_pairs:
-            self.check_extension(pser, psser._to_pandas())
-            self.check_extension(ps.from_pandas(pser), psser)
-
-    def test_isnull(self):
-        for pser, psser in self.intergral_extension_pser_psser_pairs:
-            self.assert_eq(pser.isnull(), psser.isnull())
-
-    def test_astype(self):
-        for pser, psser in self.intergral_extension_pser_psser_pairs:
-            for dtype in self.extension_dtypes:
-                if dtype in self.string_extension_dtype:
-                    self.check_extension(pser.astype(dtype), psser.astype(dtype))
-                else:
-                    self.check_extension(pser.astype(dtype), psser.astype(dtype))
-        for pser, psser in self.intergral_extension_pser_psser_pairs:
-            self.assert_eq(pser.astype(float), psser.astype(float))
-            self.assert_eq(pser.astype(np.float32), psser.astype(np.float32))
-            with ps.option_context("compute.eager_check", True):
-                self.assertRaisesRegex(
-                    ValueError,
-                    "Cannot convert integrals with missing values to bool",
-                    lambda: psser.astype(bool),
-                )
-                self.assertRaisesRegex(
-                    ValueError,
-                    "Cannot convert integrals with missing values to integer",
-                    lambda: psser.astype(int),
-                )
-                self.assertRaisesRegex(
-                    ValueError,
-                    "Cannot convert integrals with missing values to integer",
-                    lambda: psser.astype(np.int32),
-                )
-            with ps.option_context("compute.eager_check", False):
-                psser.astype(bool)
-                psser.astype(int)
-                psser.astype(np.int32)
-
-    def test_neg(self):
-        for pser, psser in self.intergral_extension_pser_psser_pairs:
-            self.check_extension(-pser, -psser)
-
-    def test_abs(self):
-        for pser, psser in self.intergral_extension_pser_psser_pairs:
-            self.check_extension(abs(pser), abs(psser))
-
-    def test_invert(self):
-        for pser, psser in self.intergral_extension_pser_psser_pairs:
-            self.check_extension(~pser, ~psser)
-
-    def test_eq(self):
-        with option_context("compute.ops_on_diff_frames", True):
-            for pser, psser in self.intergral_extension_pser_psser_pairs:
-                self.check_extension(pser == pser, (psser == psser).sort_index())
-
-    def test_ne(self):
-        with option_context("compute.ops_on_diff_frames", True):
-            for pser, psser in self.intergral_extension_pser_psser_pairs:
-                self.check_extension(pser != pser, (psser != psser).sort_index())
-
-    def test_lt(self):
-        with option_context("compute.ops_on_diff_frames", True):
-            for pser, psser in self.intergral_extension_pser_psser_pairs:
-                self.check_extension(pser < pser, (psser < psser).sort_index())
-
-    def test_le(self):
-        with option_context("compute.ops_on_diff_frames", True):
-            for pser, psser in self.intergral_extension_pser_psser_pairs:
-                self.check_extension(pser <= pser, (psser <= psser).sort_index())
-
-    def test_gt(self):
-        with option_context("compute.ops_on_diff_frames", True):
-            for pser, psser in self.intergral_extension_pser_psser_pairs:
-                self.check_extension(pser > pser, (psser > psser).sort_index())
-
-    def test_ge(self):
-        with option_context("compute.ops_on_diff_frames", True):
-            for pser, psser in self.intergral_extension_pser_psser_pairs:
-                self.check_extension(pser >= pser, (psser >= psser).sort_index())
-
-    def test_xor(self):
-        for psser in self.intergral_extension_pssers:
-            self.assertRaisesRegex(
-                TypeError,
-                "XOR can not be applied to given types.",
-                lambda: psser ^ 1,
-            )
-            self.assertRaisesRegex(
-                TypeError,
-                "XOR can not be applied to given types.",
-                lambda: psser ^ psser,
-            )
-            self.assertRaisesRegex(
-                TypeError,
-                "XOR can not be applied to given types.",
-                lambda: psser ^ False,
-            )
-
-    def test_rxor(self):
-        for psser in self.intergral_extension_pssers:
-            self.assertRaisesRegex(
-                TypeError,
-                "XOR can not be applied to given types.",
-                lambda: 1 ^ psser,
-            )
-            self.assertRaisesRegex(
-                TypeError,
-                "XOR can not be applied to given types.",
-                lambda: False ^ psser,
-            )
-
-
-@unittest.skipIf(
-    not extension_float_dtypes_available, "pandas extension float dtypes are not available"
-)
-class FractionalExtensionOpsTestsMixin:
-    @property
-    def fractional_extension_psers(self):
-        return [
-            pd.Series([0.1, 0.2, 0.3, None], dtype=dtype)
-            for dtype in self.fractional_extension_dtypes
-        ]
-
-    @property
-    def fractional_extension_pssers(self):
-        return [ps.from_pandas(pser) for pser in self.fractional_extension_psers]
-
-    @property
-    def fractional_extension_pser_psser_pairs(self):
-        return zip(self.fractional_extension_psers, self.fractional_extension_pssers)
-
-    def test_from_to_pandas(self):
-        for pser, psser in self.fractional_extension_pser_psser_pairs:
-            self.check_extension(pser, psser._to_pandas())
-            self.check_extension(ps.from_pandas(pser), psser)
-
-    def test_isnull(self):
-        for pser, psser in self.fractional_extension_pser_psser_pairs:
-            self.assert_eq(pser.isnull(), psser.isnull())
-
-    def test_astype(self):
-        for pser, psser in self.fractional_extension_pser_psser_pairs:
-            for dtype in self.extension_dtypes:
-                self.check_extension(pser.astype(dtype), psser.astype(dtype))
-        for pser, psser in self.fractional_extension_pser_psser_pairs:
-            self.assert_eq(pser.astype(float), psser.astype(float))
-            self.assert_eq(pser.astype("category"), psser.astype("category"))
-            self.assert_eq(pser.astype(np.float32), psser.astype(np.float32))
-            with ps.option_context("compute.eager_check", True):
-                self.assertRaisesRegex(
-                    ValueError,
-                    "Cannot convert fractions with missing values to bool",
-                    lambda: psser.astype(bool),
-                )
-                self.assertRaisesRegex(
-                    ValueError,
-                    "Cannot convert fractions with missing values to integer",
-                    lambda: psser.astype(int),
-                )
-                self.assertRaisesRegex(
-                    ValueError,
-                    "Cannot convert fractions with missing values to integer",
-                    lambda: psser.astype(np.int32),
-                )
-            with ps.option_context("compute.eager_check", False):
-                psser.astype(bool)
-                psser.astype(int)
-                psser.astype(np.int32)
-
-    def test_neg(self):
-        # pandas raises "TypeError: bad operand type for unary -: 'FloatingArray'"
-        for dtype in self.fractional_extension_dtypes:
-            self.assert_eq(
-                ps.Series([-0.1, -0.2, -0.3, None], dtype=dtype),
-                -ps.Series([0.1, 0.2, 0.3, None], dtype=dtype),
-            )
-
-    def test_abs(self):
-        for pser, psser in self.fractional_extension_pser_psser_pairs:
-            self.check_extension(abs(pser), abs(psser))
-
-    def test_invert(self):
-        for psser in self.fractional_extension_pssers:
-            self.assertRaises(TypeError, lambda: ~psser)
-
-    def test_eq(self):
-        with option_context("compute.ops_on_diff_frames", True):
-            for pser, psser in self.fractional_extension_pser_psser_pairs:
-                self.check_extension(pser == pser, (psser == psser).sort_index())
-
-    def test_ne(self):
-        with option_context("compute.ops_on_diff_frames", True):
-            for pser, psser in self.fractional_extension_pser_psser_pairs:
-                self.check_extension(pser != pser, (psser != psser).sort_index())
-
-    def test_lt(self):
-        with option_context("compute.ops_on_diff_frames", True):
-            for pser, psser in self.fractional_extension_pser_psser_pairs:
-                self.check_extension(pser < pser, (psser < psser).sort_index())
-
-    def test_le(self):
-        with option_context("compute.ops_on_diff_frames", True):
-            for pser, psser in self.fractional_extension_pser_psser_pairs:
-                self.check_extension(pser <= pser, (psser <= psser).sort_index())
-
-    def test_gt(self):
-        with option_context("compute.ops_on_diff_frames", True):
-            for pser, psser in self.fractional_extension_pser_psser_pairs:
-                self.check_extension(pser > pser, (psser > psser).sort_index())
-
-    def test_ge(self):
-        with option_context("compute.ops_on_diff_frames", True):
-            for pser, psser in self.fractional_extension_pser_psser_pairs:
-                self.check_extension(pser >= pser, (psser >= psser).sort_index())
-
-
 class NumOpsTests(
     NumOpsTestsMixin,
-    OpsTestBase,
-    PandasOnSparkTestCase,
-):
-    pass
-
-
-class IntegralExtensionOpsTests(
-    IntegralExtensionOpsTestsMixin,
-    OpsTestBase,
-    PandasOnSparkTestCase,
-):
-    pass
-
-
-class FractionalExtensionOpsTests(
-    FractionalExtensionOpsTestsMixin,
     OpsTestBase,
     PandasOnSparkTestCase,
 ):
