+++ python/pyspark/pandas/tests/groupby/test_split_apply.py
@@ -160,7 +160,6 @@ class GroupbySplitApplyMixin(GroupbySplitApplyTestingFuncMixin):
     def test_split_apply_combine_on_series(self):
         funcs = [
             ((True, False), ["sum"]),
-            ((True, True), ["mean"]),
         ]
         self._test_split_apply_func(funcs)
