+++ python/pyspark/pandas/tests/groupby/test_stat.py
@@ -124,37 +124,6 @@ class GroupbyStatMixin(GroupbyStatTestingFuncMixin):
             expected_error=ValueError if using_pandas3 else None,
         )
 
-    def test_median(self):
-        psdf = ps.DataFrame(
-            {
-                "a": [1.0, 1.0, 1.0, 1.0, 2.0, 2.0, 2.0, 3.0, 3.0, 3.0],
-                "b": [2.0, 3.0, 1.0, 4.0, 6.0, 9.0, 8.0, 10.0, 7.0, 5.0],
-                "c": [3.0, 5.0, 2.0, 5.0, 1.0, 2.0, 6.0, 4.0, 3.0, 6.0],
-            },
-            columns=["a", "b", "c"],
-            index=[7, 2, 4, 1, 3, 4, 9, 10, 5, 6],
-        )
-        # DataFrame
-        expected_result = ps.DataFrame(
-            {"b": [2.0, 8.0, 7.0], "c": [3.0, 2.0, 4.0]}, index=pd.Index([1.0, 2.0, 3.0], name="a")
-        )
-        self.assert_eq(expected_result, psdf.groupby("a").median().sort_index())
-        # Series
-        expected_result = ps.Series(
-            [2.0, 8.0, 7.0], name="b", index=pd.Index([1.0, 2.0, 3.0], name="a")
-        )
-        self.assert_eq(expected_result, psdf.groupby("a")["b"].median().sort_index())
-
-        with self.assertRaisesRegex(TypeError, "accuracy must be an integer; however"):
-            psdf.groupby("a").median(accuracy="a")
-
-        if LooseVersion(pd.__version__) >= "3.0.0":
-            # pandas < 3 raises an error when numeric_only is False or None
-            self._test_stat_func(
-                lambda groupby_obj: groupby_obj.median(numeric_only=None),
-                expected_error=ValueError if using_pandas3 else None,
-            )
-
 
 class GroupbyStatTests(
     GroupbyStatMixin,
