+++ python/pyspark/sql/tests/connect/arrow/test_parity_arrow_cogrouped_map_misc.py
@@ -15,18 +15,12 @@
 # limitations under the License.
 #
 
-from pyspark.pandas.tests.data_type_ops.test_num_ops import NumOpsTestsMixin
-from pyspark.pandas.tests.data_type_ops.testing_utils import OpsTestBase
-from pyspark.testing.pandasutils import PandasOnSparkTestUtils
+
+from pyspark.sql.tests.arrow.test_arrow_cogrouped_map_misc import CogroupedMapInArrowMiscTestsMixin
 from pyspark.testing.connectutils import ReusedConnectTestCase
 
 
-class NumOpsParityTests(
-    NumOpsTestsMixin,
-    PandasOnSparkTestUtils,
-    OpsTestBase,
-    ReusedConnectTestCase,
-):
+class CogroupedMapInArrowMiscParityTests(CogroupedMapInArrowMiscTestsMixin, ReusedConnectTestCase):
     pass
